refactor: add UserService
