+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -1,9 +1,12 @@
 package ru.yandex.practicum.filmorate.controller;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.service.UserService;
 
 import javax.validation.Valid;
 import java.util.HashMap;
@@ -18,80 +21,29 @@ import static ru.yandex.practicum.filmorate.constans.Settings.BAN_LIST_FIND_LOGI
 @RequestMapping("/users")
 @Slf4j
 public class UserController {
-    private final Map<Integer, User> users = new HashMap<>();
-    private int idGenerator = 1;
+    private UserService userService;
+
+    @Autowired
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
 
     @GetMapping
     public List<User> getAllUsers() {
-        log.info("Сделан запрос на получение списка всех пользователей. *Работает фильтр BanListFindLogin.properties");
-        return users.values().stream().filter(user -> !BAN_LIST_FIND_LOGIN.contains(user.getLogin()))
-                .collect(Collectors.toList());
+        log.debug("Сделан запрос на получение списка всех пользователей");
+        return userService.getAllUsers();
     }
 
     @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
     public User createUser(@Valid @RequestBody User user) {
-        if (isCheckLoginInBanList(user.getLogin())) {
-            throw new ValidationException("Регистрировать пользователя с такими именем запрещено - " + user.getLogin());
-        }
-        if (isCheckLoginOnDuplicate(user.getLogin())) {
-            throw new ValidationException("Пользователь с таким же логином уже имеется в системе - " + user.getLogin());
-        }
-        if (isCheckEmailInDateBase(user.getEmail())) {
-            throw new ValidationException("Пользователь с таким email - " + user.getEmail() + " уже существует");
-        }
-        if (isCheckName(user)) {
-            user.setName(user.getLogin());
-        }
-        user.setId(idGenerator++);
-        users.put(user.getId(), user);
-        log.info("В БД успешно добавлен пользователь с ID={}", user.getId());
-        return user;
+        log.debug("Сделан запрос на добавление нового пользователя");
+        return userService.createUser(user);
     }
 
     @PutMapping
     public User updateUser(@Valid @RequestBody User user) {
-        int userId = user.getId();
-        if (users.containsKey(userId)) {
-            if (isCheckLoginInBanList(user.getLogin())) {
-                throw new ValidationException("Изменение на такой \"login\" - " + user.getLogin() + " запрещено");
-            }
-            if (isCheckEmailInDateBase(user.getEmail())) {
-                throw new ValidationException("Данный email - " + user.getEmail() + " уже находится в БД");
-            }
-            User oldUser = users.get(userId);
-            if (isCheckName(user)) {
-                user.setName(user.getLogin());
-            }
-            users.put(userId, user);
-            if (user.equals(oldUser)) {
-                log.warn("При обновлении данных аккаунта, пользователь с ID={} не дал новых данных " +
-                        "если это сообщение повторится, на это стоит обратить внимание", userId);
-            }
-            log.info("Пользователь с ID={} успешно обновлен!\n Старый аккаунт: {},\n Новый аккаунт: {}",
-                    userId, oldUser.toString(), users.get(userId));
-            return user;
-        } else {
-            throw new ValidationException("Пользователя с таким ID=" + userId + " не существует");
-        }
-    }
-
-    private Boolean isCheckName(User user) {
-        if (user.getName() == null || user.getName().isBlank()) {
-            log.info("Пользователь не указал имени. Поле \"name\" берется из логина - {}", user.getLogin());
-            return true;
-        }
-        return false;
-    }
-
-    private Boolean isCheckEmailInDateBase(String emailCheck) {
-        return users.values().stream().anyMatch(user -> user.getEmail().equalsIgnoreCase(emailCheck));
-    }
-
-    private Boolean isCheckLoginInBanList(String login) {
-        return BAN_LIST_ADD_LOGIN.stream().anyMatch(login::equalsIgnoreCase);
-    }
-
-    private Boolean isCheckLoginOnDuplicate(String login) {
-        return users.values().stream().anyMatch(user -> user.getLogin().equalsIgnoreCase(login));
+        log.debug("Сделан запрос на обновление пользователя с ID={}", user.getId());
+        return userService.updateUser(user);
     }
 }
